Refactor mouse input to account for decimals

+++ src/inputs/Inputs.h
@@ -18,7 +18,7 @@ namespace Inputs
     /*
      * Moves the mouse cursor by the specified amount.
      */
-    void MoveMouseRaw(int dx, int dy);
+    void MoveMouseRaw(float dx, float dy);
 
     /*
      * Returns true if a controller is connected.
